+++ beltradar/__init__.py
@@ -1,6 +1,6 @@
 """Initialize the app"""
 
-__version__ = "0.0.8"
+__version__ = "0.0.9"
 __title__ = "Belt Radar"
 
 __package_name__ = "aa-beltradar"